Refactor cliconfig so that the default constructor can exist in the package.
