+++ cliconfig/credentials/default_store.go
@@ -3,12 +3,12 @@ package credentials
 import (
 	"os/exec"
 
-	"github.com/docker/docker/cliconfig"
+	"github.com/docker/docker/cliconfig/configfile"
 )
 
 // DetectDefaultStore sets the default credentials store
 // if the host includes the default store helper program.
-func DetectDefaultStore(c *cliconfig.ConfigFile) {
+func DetectDefaultStore(c *configfile.ConfigFile) {
 	if c.CredentialsStore != "" {
 		// user defined
 		return
